small fixes

+++ helpers/users.py
@@ -8,8 +8,11 @@ headers = {"accept": "application/json"}
 def get_user_by_id(user_id):
 	response = test_api().get("users/" + user_id, headers=headers)
 	result = json.loads(response.text)
-	assert response.status_code == 200
-	assert str(result[0]['id']) == user_id
+	if response.status_code == 200:
+		assert str(result[0]['id']) == user_id
+		return result[0]['id']
+	else:
+		assert response.status_code == 404
 
 
 def create_user(name, email, age):
@@ -18,6 +21,7 @@ def create_user(name, email, age):
 	result = json.loads(response.text)
 	assert response.status_code == 200
 	assert str(result['data']['name']) == name
+
 
 #for example update user name
 def update_user_by_id(user_id, name):
@@ -29,13 +33,16 @@ def update_user_by_id(user_id, name):
 
 
 def delete_user_by_id(user_id):
-	response = test_api().delete("users/" + user_id,  headers=headers)
-	result = json.loads(response.text)
-	assert response.status_code == 200
-	assert str(result['data']['id']) == user_id
+	result = get_user_by_id(user_id)
+	if result:
+		response = test_api().delete("users/" + user_id,  headers=headers)
+		result = json.loads(response.text)
+		assert response.status_code == 200
+		assert str(result['data']['id']) == user_id
+	else:
+		assert result == None
 
 
 def get_users():
-	response = test_api().get("users/", headers=headers)
-	result = json.loads(response.text)
+	response = test_api().get("users", headers=headers)
 	assert response.status_code == 200
